fix: extract shared e2e helpers and fix remaining flaky keyboard tests
Move waitForEmailListReady and pressKeyUntilVisible to launch-helpers.ts
for reuse across test files. Fix flaky "Escape clears multi-selection"
(keyboard-flow:632) and "full view shows sender name" (sender-profile:226)
by replacing fixed waits with retry-based key pressing.
